wip / almost done with Product model, just needs an update method for updating the picked_up boolean.
Also, I changed the format of product prices to a string that can be converted later to keep the decimal (which is removed when JSON formatted).
I changed the `pref_zip` field on the Shopper model to a ForeignKey linked field to Location called `current_store`.
This field can be `null` and I have set it to _always_ be `null` when a Shopper is first registered.
A PUT request will be made when the Shopper decides to select a current store.