create, list and retrieve products. Update not necessary.
I decided not to write code for updating a product, but instead removed the `picked_up` field since multiple shoppers should technically be able to use the same product. So, I kept the `completed` field on the `ShoppingList` model, since that will be specific to a product, location, and shopper whenthe `completed` field is updated. Ideally, I would love to check if the product is already in my database by querying for the `kroger_id`, but I would need to also have the location_id in the query and on the Product model since aisle locations vary by store. Going to put that one on the back burner for now.